fix: combine 2 modules to one (create util.js)
